requirements.txt
flask==2.2.0
torch==2.2.0
torchvision==0.17.0
pillow==9.4.0
numpy==1.24.0
setuptools==68.0.0